[feat] 파티 어드민 페이지 수정사항 변경, 템플릿 어드민 모달창 연결
- 파티 어드민 페이지 수정사항 변경
- 템플릿 어드민 모달창 연결
