+++ components/party/PartyCategory.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import {
   Paper,
   Table,
@@ -8,10 +8,9 @@ import {
   TableRow,
 } from '@mui/material';
 import { PartyCategory } from 'types/partyTypes';
-import { mockInstance } from 'utils/mockAxios';
 import { tableFormat } from 'constants/admin/table';
 import { AdminTableHead } from 'components/admin/common/AdminTable';
-import usePartyRoom from 'hooks/party/usePartyList';
+import usePartyCategory from 'hooks/party/usePartyCategory';
 import styles from 'styles/party/PartyMain.module.scss';
 
 const tableTitle: { [key: string]: string } = {
@@ -21,64 +20,42 @@ const tableTitle: { [key: string]: string } = {
 };
 
 export default function PartyCategories() {
-  // const { categorys } = usePartyRoom();
+  const { categorys, createCategory, deleteCategory } = usePartyCategory();
   const [newCategoryName, setNewCategoryName] = useState('');
-  // usecategory hook사용
-  // const categories: PartyCategory[] = categorys.map((cate) => ({
-  //   categoryId: cate.categoryId,
-  //   categoryName:
-  //     categorys.find((c) => c.categoryId === cate.categoryId)?.categoryName ??
-  //     '???',
-  // }));
-
-  // useEffect(() => {
-  //   console.log('render');
-  // }, [categorys]);
 
   const handleConfirm = () => {
     if (newCategoryName.trim() !== '') {
-      const newCategory = {
-        categoryName: newCategoryName,
-      };
-      mockInstance
-        .post('/party/admin/categorys', newCategory)
-        .then(() => {
-          setNewCategoryName('');
-        })
-        .catch((error) => {
-          console.error('카테고리 추가 중 오류가 발생했습니다:', error);
-        });
+      createCategory(newCategoryName);
     }
   };
 
-  const deleteCategory = (categoryId: number) => {
-    mockInstance
-      .delete(`/party/admin/categorys/${categoryId}`)
-      .catch((error) => {
-        console.error('카테고리 삭제 중 오류가 발생했습니다:', error);
-      });
+  const deletehandler = (categoryId: number) => {
+    deleteCategory(categoryId);
   };
-  console.log(
-    tableFormat['partyCategory'].columns.map(
-      (columnName) => tableTitle[columnName]
-    )
-  );
+
   return (
     <div className={styles.userManagementWrap}>
       <div className={styles.header}>
         <span className={styles.title}>카테고리 관리</span>
-        <input
-          type='text'
-          value={newCategoryName}
-          onChange={(e) => setNewCategoryName(e.target.value)}
-        />
-        <button onClick={handleConfirm}>확인</button>
+        <div>
+          <input
+            type='text'
+            value={newCategoryName}
+            onChange={(e) => setNewCategoryName(e.target.value)}
+          />
+          <button
+            onClick={handleConfirm}
+            className={`${styles.button_1} ${styles.add}`}
+          >
+            추가
+          </button>
+        </div>
       </div>
       <TableContainer className={styles.tableContainer} component={Paper}>
         <Table className={styles.table} aria-label='UserManagementTable'>
           <AdminTableHead tableName={'partyCategory'} table={tableTitle} />
           <TableBody className={styles.tableBody}>
-            {/* {categories.map((c) => (
+            {categorys.map((c) => (
               <TableRow key={c.categoryId} className={styles.tableRow}>
                 {tableFormat['partyCategory'].columns.map((columnName) => {
                   return (
@@ -87,7 +64,10 @@ export default function PartyCategories() {
                       key={columnName}
                     >
                       {columnName === 'delete' ? (
-                        <button onClick={() => deleteCategory(c.categoryId)}>
+                        <button
+                          onClick={() => deletehandler(c.categoryId)}
+                          className={`${styles.button_1} ${styles.delete}`}
+                        >
                           삭제
                         </button>
                       ) : (
@@ -97,7 +77,7 @@ export default function PartyCategories() {
                   );
                 })}
               </TableRow>
-            ))} */}
+            ))}
           </TableBody>
         </Table>
       </TableContainer>
