+++ styles/party/PartyMain.module.scss
@@ -19,6 +19,51 @@ $light-green: #67a64a;
   }
 }
 
+.userManagementWrap {
+  width: 100%;
+  height: 100%;
+
+  .header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 95%;
+    min-width: 710px;
+    margin: 10px auto;
+    .title {
+      font-size: $big-font;
+      font-weight: 600;
+      line-height: 150%;
+    }
+  }
+
+  .button_1 {
+    padding: 4px 12px;
+    margin-bottom: 10px;
+    margin-left: 30px;
+    font-size: $small-font;
+    cursor: pointer;
+    border: none;
+    border-radius: $medium-radius;
+  }
+  .delete {
+    color: #ffffff;
+    background-color: #d73535;
+    &:hover {
+      background-color: #d71e1e;
+      transition: 0.3s;
+    }
+  }
+  .add {
+    color: white;
+    background-color: #10b981;
+    &:hover {
+      background-color: #059669;
+      transition: 0.3s;
+    }
+  }
+}
+
 .categoryNav {
   ul {
     display: flex;
