+++ types/modalTypes.ts
@@ -13,6 +13,7 @@ import { StoreManualMode } from 'types/storeTypes';
 import { ICoin } from 'types/userTypes';
 import { ITournament } from './admin/adminTournamentTypes';
 import { GameMode } from './gameTypes';
+import { PartyGameTemplate } from './partyTypes';
 import { TournamentInfo } from './tournamentTypes';
 
 type EventModal = 'WELCOME' | 'ANNOUNCEMENT';
@@ -55,7 +56,8 @@ type AdminModal =
   | 'CHECK_SEND_NOTI'
   | 'TOURNAMENT_BRAKET_EDIT'
   | 'TOURNAMENT_PARTICIPANT_EDIT'
-  | 'PARTY_EDIT';
+  | 'PARTY_EDIT'
+  | 'PARTY_TEMPLATE';
 
 type ModalName =
   | null
@@ -153,4 +155,5 @@ export interface Modal {
   tournament?: ITournament;
   tournamentId?: number;
   partyReport?: PartyReportModalData;
+  template?: PartyGameTemplate | null;
 }
